Cleanup + 27

+++ 009.c
@@ -25,7 +25,7 @@ int main() {
 		printf("The product abc is %ld\n", (long)(a * b * c));
 		printf("a is %d\n", a);
 		printf("b is %d\n", b);
-		printf("c is %f\n", c);
+		printf("c is %d\n", (int)c);
 	}
 	else {
 		printf("No combination of natural numbers found\n");
